Separate Report from SyncReport
